+++ App/RefinerServer/RefinerServer/ImageRepresentation.h
@@ -15,30 +15,30 @@
 	{
 	public:
 		ImageRepresentation(std::string filename,
-			Eigen::Matrix4d pCameraViewTransform,
-			Eigen::Matrix4d pCameraProjectionTransform);
+			Eigen::Matrix4f pCameraViewTransform,
+			Eigen::Matrix4f pCameraProjectionTransform);
 
 		/// <summary>
 		/// compute 3D projection of 2D point "pixel" onto surface defined by normal "surface_normal" and vertex "vertex"
 		/// </summary>
-		Eigen::Vector3d project2dto3d(Eigen::Vector3d surface_normal, Eigen::Vector3d vertex, cv::Point2d p);
+		Eigen::Vector3f project2dto3d(Eigen::Vector3f surface_normal, Eigen::Vector3f vertex, cv::Point2f p);
 		
 		/// <summary>
 		/// Compute corrsponding patch in camera 2 given the patch in camera 1 (via reprojection and projective unwarping)
 		/// </summary>
-		double computeDistortedPatchCorrelation(ImageRepresentation& image2, Eigen::Vector3d surface_normal, Eigen::Vector3d vertex, cv::Size patch_size, int colorFlag);
+		float computeDistortedPatchCorrelation(ImageRepresentation& image2, Eigen::Vector3f surface_normal, Eigen::Vector3f vertex, cv::Size patch_size, int colorFlag);
 		
 
 		/// <summary>
 		/// Compute corrsponding patch in camera 2 given the patch in camera 1 (without unwarping)
 		/// </summary>
-		double computePatchCorrelation(ImageRepresentation& image2, Eigen::Vector3d surface_normal, Eigen::Vector3d vertex, cv::Size patch_size);
+		float computePatchCorrelation(ImageRepresentation& image2, Eigen::Vector3f surface_normal, Eigen::Vector3f vertex, cv::Size patch_size);
 
 
 		/// <summary>
 		/// Compute view quality
 		/// </summary>
-		double getViewQuality(Eigen::Vector3d vertex, Eigen::Vector3d normal, ImageRepresentation& image2);
+		float getViewQuality(Eigen::Vector3f vertex, Eigen::Vector3f normal, ImageRepresentation& image2);
 
 
 		const std::string filename;
@@ -46,6 +46,6 @@
 		cv::Mat ocvImage;
 		unsigned width, height;
 		int x_size, y_size;
-		Eigen::Matrix4d CameraViewTransform;
-		Eigen::Matrix4d CameraProjectionTransform;
+		Eigen::Matrix4f CameraViewTransform;
+		Eigen::Matrix4f CameraProjectionTransform;
 	};
